Move MatchDuplicateDTO to a new file
