+++ API/DTOs/MatchDuplicateCollectionDTO.cs
@@ -2,15 +2,6 @@
 
 namespace API.DTOs;
 
-public class MatchDuplicateDTO
-{
-	public long OsuMatchId { get; set; }
-	public string Name { get; set; } = string.Empty!;
-	public bool? VerifiedAsDuplicate { get; set; }
-	public string? VerifiedByUsername { get; set; }
-	public int? VerifiedByUserId { get; set; }
-}
-
 public class MatchDuplicateCollectionDTO
 {
 	public int Id { get; set; }
